Working with --source

+++ tools/project_api.py
@@ -2,7 +2,7 @@
 import sys
 from os.path import join, abspath, dirname, exists
 from os.path import basename, relpath, normpath
-from os import mkdir
+from os import mkdir, remove
 ROOT = abspath(join(dirname(__file__), ".."))
 sys.path.insert(0, ROOT)
 
@@ -59,13 +59,11 @@ def subtract_basepath(resources, export_path):
                                                      export_path))
 
 
-def export_project(src_paths, export_path, target, ide,
-                   libraries_paths=None, options=None, linker_script=None,
-                   clean=False, notify=None, verbose=False, name=None,
-                   inc_dirs=None, jobs=1, silent=False,
-                   report=None, properties=None, project_id=None,
-                   project_description=None, extra_verbose=False, config=None,
-                   build=False, macros=[]):
+def prepare_project(src_paths, export_path, target, ide,
+                    libraries_paths=None, options=None, linker_script=None,
+                    clean=False, notify=None, verbose=False, name=None,
+                    inc_dirs=None, jobs=1, silent=False, extra_verbose=False,
+                    config=None, macros=None):
     """ This function builds a project. Project can be for example one test / UT
     """
 
@@ -77,10 +75,10 @@ def export_project(src_paths, export_path, target, ide,
         src_paths.extend(libraries_paths)
 
     # Export Directory
-    if clean:
-        if exists(export_path):
-            rmtree(export_path)
-    mkdir(export_path)
+    if exists(export_path) and clean:
+        rmtree(export_path)
+    if not exists(export_path):
+        mkdir(export_path)
 
     exporter_cls, toolchain_name = get_exporter_toolchain(ide)
 
@@ -96,67 +94,29 @@ def export_project(src_paths, export_path, target, ide,
     if name is None:
         name = basename(normpath(abspath(src_paths[0])))
 
-    # Initialize reporting
-    if report != None:
-        start = time()
-        # If project_id is specified, use that over the default name
-        id_name = project_id.upper() if project_id else name.upper()
-        description = project_description if project_description else name
-        vendor_label = target.extra_labels[0]
-        prep_report(report, target.name, toolchain_name, id_name)
-        cur_result = create_result(target.name, toolchain_name, id_name,
-                                   description)
-        if properties != None:
-            prep_properties(properties, target.name, toolchain_name,
-                            vendor_label)
 
-    try:
-        # Call unified scan_resources
-        resources = scan_resources(src_paths, toolchain, inc_dirs=inc_dirs)
-        toolchain.build_dir = export_path
-        config_header = toolchain.get_config_header()
-        resources.headers.append(config_header)
-        resources.file_basepath[config_header] = dirname(config_header)
+    # Call unified scan_resources
+    resources = scan_resources(src_paths, toolchain, inc_dirs=inc_dirs)
+    toolchain.build_dir = export_path
+    config_header = toolchain.get_config_header()
+    resources.headers.append(config_header)
+    resources.file_basepath[config_header] = dirname(config_header)
 
-        # Change linker script if specified
-        if linker_script is not None:
-            resources.linker_script = linker_script
+    # Change linker script if specified
+    if linker_script is not None:
+        resources.linker_script = linker_script
+
+    return resources, toolchain
 
 
-        temp = copy.deepcopy(resources)
-        subtract_basepath(resources, export_path)
+def export_project(resources, export_path, target, name, toolchain, ide, macros=None):
+        exporter_cls, toolchain_name = get_exporter_toolchain(ide)
         exporter = exporter_cls(target, export_path, name, toolchain,
                                 extra_symbols=macros, resources=resources)
         exporter.generate()
         files = exporter.generated_files
 
-        if report != None:
-            end = time()
-            cur_result["elapsed_time"] = end - start
-            cur_result["output"] = toolchain.get_output()
-            cur_result["result"] = "OK"
-            cur_result["memory_usage"] = toolchain.map_outputs
-
-            add_result_to_report(report, cur_result)
-
-        return files, temp, export_path, name
-
-    except Exception:
-        if report != None:
-            end = time()
-
-            cur_result["result"] = "FAIL"
-
-            cur_result["elapsed_time"] = end - start
-
-            toolchain_output = toolchain.get_output()
-            if toolchain_output:
-                cur_result["output"] += toolchain_output
-
-            add_result_to_report(report, cur_result)
-
-        # Let Exception propagate
-        raise
+        return files
 
 def zip_export(file_name, prefix, resources, project_files):
     """Create a zip file from an exported project.
